pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zenrube"
version = "1.0.0"
description = "Modular Cognitive Platform for Expert AI orchestration and self-publishing to Rube.app."
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
  {name = "Vlad F", email = "vladinc@gmail.com"}
]
keywords = ["AI", "automation", "Rube.app", "Zenrube", "expert-system"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent"
]
dependencies = [
  "pydantic>=1.10",
  "PyYAML>=6.0"
]

[project.urls]
Homepage = "https://github.com/vmanoilov/zenrube-mcp"
Repository = "https://github.com/vmanoilov/zenrube-mcp"

[tool.setuptools]
package-dir = {"" = "."}
packages = ["zenrube", "zenrube.experts"]
include-package-data = true

[project.scripts]
zenrube = "zenrube.cli:main"
